requirements.txt
mcp>=1.0.0
databricks-sdk>=0.20.0
psycopg2-binary>=2.9.0
